feat(plugin) separate validation errors and fails
Add a new string return value to PluginService.Validate(). When the
response is a validation failure, this string is the response body
message and the error is nil. Actual errors (e.g. connection timeouts,
response parse failures, etc.) still set the error.

Previously, failures also set the error. Downstream applications were
unable to determine whether the validation attempt was unsuccessful or
if the validation attempt was successful and confirmed the payload was
invalid.

Add tests for Validate().