feat: improve handling zero-value fields with defaults
We recently forked and patched mergo (https://github.com/Kong/mergo/pull/1)
to improve the way we can handle zero-value fields with defaults values.
Currently, if an entity field is set with a zero-value (e.g. a target
with weight=0, a route with strip_path=false etc.), we will overwrite it
when merging it with its default if this is not a zero-value itself.

This update the mergo dependence and add a custom transformer to support
such cases as well. Also, add few test cases with zero-value fields.