fix: only fills defaults for plugins config schemas
A few issues ([here](https://github.com/Kong/deck/issues/612) and
[here](https://github.com/Kong/deck/issues/610)) have been opened
on decK that suggest we sometimes parse wrongly the plugins defaults
coming from their schema, merging into the config object fields
that are not part of the config schema itself.

This makes sure that we only merge together the right attributes.