fix(tests) skip some tests on Enterprise
Skip the consumer list and credential list on Enterprise. The presence
of admins (which wrap hidden consumers and credentials) affects those
resources' pagination. The tests require very specific pagination
behavior based on the set of consumers they create, and we cannot
isolate the admin in its own test, as it's used in all Enterprise tests.