feat: support both POST and PUT for target entities
Starting with 3.0, POSTs are not supported anymore for the '/targets'
endpoints and they will be replaced by PUTs. This means that we need
a way to support both methods in order to offer backward compatibility.

This introduces a 'retry' mechanism to be leveraged in the case of
POST returning 405s.