fix: correctly ingest entity defaults with maps values
The current implementation of the `fillConfigRecord` function
works as follows:
- navigate the entity structure
- for each field, check whether it's set or not
- if it's already set, move onto the next field
- if it's not set, check the entity schema and set a default
  if that's defined

This works well in most cases, except for plugins fields being
maps. For such fields, this simple check may not be enough since
the field may be composed of several sub-fields possibly
having their own default values. This means that moving onto the
next field when the current map field is set may results into
its inner sub-fields not being set with default values.

In decK, where this function is heavily used, this issue results
in misleading deployment diffs.

This commit expands the check logic to also verify that all
default values are set, even those of maps fields.