fix: fill config values and defaults in shorthand fields (#458)
* fix: fill config values and defaults in shorthand fields

As of now, deprecated fields are renamed to shorthand_fields in all
schemas, as per the conventions. Till now, we do not fill any
values for these fields while attempting to fill config records for
plugins. This creates a visualisation problem in decK. Everytime, a
deck diff or deck sync command is issued, it shows that the deprecated
field values are changed and need to be updated, thus running an
unnecessary update process each time and also confusing end users.
Check #1251: https://github.com/Kong/deck/issues/1251 for clarity.

This fix attempts to fill defaults for the shorthand_fields, retaining
their values, if passed. Also, since shorthand_fields take priority
over normal/nested fields in the gateway, if any changes are detected in
shorthand_fields, it is backfilled to nested fields as well.

* test: unit test added for shorthand_fields filling

* chore: godoc for new functions added, comments added for clarification

* chore: typo correction and preallocation of slice for backward path translation