fix(utils): support auto fields
When fields marked as `auto` are passed to kong without a value
they are auto generated by the gateway. In that case, comparing with
"defaults" doesn't work, so in order to compute a correct diff they must
be copied from Kong's configuration.

This commit addS a new FillPluginsDefaultsAutoFields utils function that
in addition to default fields, adds auto fields to the configuration.
It takes an additional parameter: oldConfig, which represents the config
from Kong: auto fields are just copied from it so that they don't show
up as differences in the diff.