fix: use replaced_with to check for deprecated fields
In schemas, shorthand_fields that are deprecated
will use "replaced_with" going forward to convey the
new fields that would be used in place of the old ones.
Earlier, "translate_backwards" was used for the same.

This change ensures that both of the above mentioned fields
can be used to find the new path that leads to the new values
for comparison purposes. This will help tools like deck
to show diff between syncs.