feat: do not list consumers in a GET Consumer Group call (#487)
* feat: do not list consumers in a GET Consumer Group call

Gateways before 3.9 return a list of consumers
along with consumer_group information, when
performing a GET call on a specific consumer_group.
This can become slow in case a consumer_group has
a lots of consumers. A separate endpoint and relevant
constructs already exist to fetch consumers for a
particular group. Thus, we are trying to not list
consumers while fetching a consumer_group by passing
a query parameter in the API call. Gateway 3.9 supports
this.

This will make decK faster for cases where a consumer-
group holds thousands of consumers.
For: https://github.com/Kong/deck/issues/1457

* chore: added PR suggestions