fix: added new method to fetch a consumer-group without consumers.
Reverted the existing `Get` method to keep listing
consumers and added a new one that uses the query
parameter `list_consumers=false` to skip listing
consumers for performance reasons.
This is done so as to fix dump errors coming in
deck when using gw 3.9+ where consumer-groups
do not show the associated consumers. We will
push the performant option behind a flag.

For: https://github.com/Kong/deck/issues/1483